create timer repository

+++ api/app/repository/task.py
@@ -1,5 +1,5 @@
 from app.core.extensions import db
-from app.repository.database_abc import DeleteDB, UpdateDB, SearchDB
+from app.repository.database_abc import DeleteDB, UpdateDB, SearchDB, SearchCategoryDB
 from sqlalchemy.exc import SQLAlchemyError
 from app.model.task_model import TaskModel
         
@@ -7,6 +7,10 @@ class DeleteTask(DeleteDB):
     def delete_db(self, id):
         try:
             task = TaskModel.query.filter(TaskModel.id == id).first()
+
+            if task is None:
+                raise ValueError('Not found')
+
             db.session.delete(task)
             db.session.commit()
         except SQLAlchemyError:
@@ -30,7 +34,7 @@ class UpdateTask(UpdateDB):
             db.session.rollback()
             raise SQLAlchemyError
         
-class SearchByID(SearchDB):
+class SearchTask(SearchDB):
     def search_db_by_id(self, id):
         try:
             task = TaskModel.query.filter(TaskModel.id == id).first()
@@ -47,6 +51,7 @@ class SearchByID(SearchDB):
         except SQLAlchemyError:
             raise SQLAlchemyError
         
+class SearchTaskCategory(SearchCategoryDB):
     def search_db_by_category_all(self, category_id):
         try:
             all_task = TaskModel.query.filter(TaskModel.category_id == category_id).all()
